cloud-only migration + enterprise hardening (non-Goose lane)
Remove the Omega/computer-use lane and the local-model (MLX/GGUF) inference
stack; make Epistemos cloud-only for LLM inference. Then deep-harden the kept
surfaces (meetings, research/agent, browsing, editor) for enterprise + MAS.

- Cloud-only removal: delete vmlx-swift + GGUFRuntimeBridge packages, the
  Omega/computer-use lane + omega-ax crate; InferenceState/TriageService
  rewritten cloud-only (0 .localMLX producers); entitlements scrubbed
  (unsigned-exec-mem, apple-events, accessibility).
- Hardening (18 findings, all CRITICAL/HIGH fixed + build-verified):
  MEET-1/4/7, CONC-2/4/13/15, SEC-3/4, BUP-1/2, WEB-1/2/3/5, BRW-1/2.
  Regression tests: BUP-1, MEET-7, CONC-4.
- Removal test-debt reconciled: 34 source-guard assertions across 10 test
  files; fixed a build-script MLX-patch removal gap (scripts/xcodebuild_epistemos.sh).
- Docs: HARDENING_AUDIT.md, PROGRESS.md, OWNER_ACTIONS.md, THREAT_MODEL.md.

Excludes concurrent-lane Goose files + RootView.swift (not this lane's work).

+++ LocalPackages/KokoroPipeline/Sources/KokoroPipeline/KokoroSynthesisExecutor.swift
@@ -651,7 +651,9 @@ private func zero(_ array: MLMultiArray) {
     case .float32:
         byteCount = array.count * MemoryLayout<Float>.size
     case .float16:
-        byteCount = array.count * MemoryLayout<Float16>.size
+        // UInt16 stands in for Float16's 2-byte size; Swift's Float16 type is
+        // unavailable on x86_64 macOS and only the byte width matters here.
+        byteCount = array.count * MemoryLayout<UInt16>.size
     case .double:
         byteCount = array.count * MemoryLayout<Double>.size
     case .int32:
